+++ CHANGELOG.md
@@ -1,1 +1,4 @@
 # v2.11.9
+
+* changed `has_p` to `has`, deprecated `has_p`
+